django_bot/requirements.txt
aio-pika==9.4.0
aioamqp==0.15.0
aioredis==2.0.1
aiormq==6.8.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
asyncstdlib==3.12.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.8
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
multidict==6.0.5
numpy==1.26.4
pamqp==3.3.0
pandas==2.2.1
pika==1.3.2
Pillow==10.0.0
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.8
pytz==2024.1
redis==5.0.2
requests==2.31.0
sentry-sdk==1.41.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.0
yarl==1.9.4
uvicorn
gunicorn
uvicorn-worker
pydub
